Simplify resolver API
Because it was much more performant to resolve all jars in a single
gradle task (vs. N jars in N tasks), the resolver API took the shape
that it's currently in - taking in a list of coordinates and returning a
list of jars.

Since we stopped using gradle resolver, the resolver API can now take a
single coordinate and output a single path (if jar is resolved). This
makes the API much easier to use in list-comprehension functions.

In addition:
- Replaced usages of AppInspectorJar with Path because Path is a lot
easier to use and AppInspectorJar should only be used when interfacing
with transport pipeline code.

Test: covered by existing tests.
Bug: 165319681
Change-Id: I828c5f40f6c71aa1241686d1deae4cca0bfc7172