[Dsl] allow adding infix properties to plugins
At least, when the existing plugin is declared in a plugins block.  It
might not do anything close to the right thing if the plugin is declared
in one of the various older ways of doing that (in an apply block, or an
apply function call with a plugin argument).

Define LiteralToInfixTransform and InfixPropertyTransform (somewhat
mirroring SingleArgToMapTransform and MapMethodTransform), which are
respectively responsible for converting a plugin declaration from
Literal to Infix form, and for looking up properties from an infix
form.  This allows deleting the need to keep around nullable Dsl
elements for apply and version, and removes a fair quantity of
nullability bugs in the temporary implementation.

Bug: 165576187
Test: added
Change-Id: Ifde16ab1b01a2a7151df56a9f3380c8c3428ae12