Handle database id not found  when fetching schema
Since dbs are opened and closed on the device, but we ask for
the schema from studio, as soon as a db is opened, race
conditions can happen. Like:

1. device sends "db open"
2. device sends "db closed"
3. studio asks for db schema

To avoid this from happening we should handle the case
in which we try to get the schema of a db that is no
there.

Test: DatabaseInspectorControllerTest.kt
Bug: 170680832
Change-Id: Iede0c53a596111d684de6c19ae74e613e91b8898