Add layout inspector features to SelectProcessAction
In a future commit, we're going to replace SelectProcessAction in layout
inspector with the one from app inspection. However, there are a handful
of issues currently that need to be resolved first:

* The "process" label needs to be more compact, because the layout
inspector currently doesn't have a lot of horizontal space

* The default stop text doesn't make sense for layout inspector, so it
needs to be updatable

* Layout inspector doesn't (currently) support offline mode, so it needs
to support filtering out detached processes.

Bug: 174765561
Test: Updated
Change-Id: I4c916e812e1279570a0ee1114c8828357149ec16