Increase timeout in HostPanelTest
HostPanel.init does a ReferencesSearch that can take a long time.
Increase the timeout in HostPanelTest.testDumbMode where we wait for
that.
I notice that the ReferencesSearch is happening in a ReadAction, and
while it's executed under progress, it's using an EmptyProgressIndicator
that isn't cancellable. We need to be more careful about executing
potentially long-running read actions, allowing them to be cancelled and
restarted if necessary. Filed 175919583 for that.

Test: this is a test fix
Fixes: 175796226
Change-Id: I325459caa15d506605e92c13b67740db206d262d