Android: fix SOE in AndroidProcessHandler#toString
AndroidProcessHandler#toString delegates to UserDataHolderBase#toString, and contains itself as a value in the CopyableUserData (see `init` block: `putCopyableUserData(SwappableProcessHandler.EXTENSION_KEY, this)`). As a result, UserDataHolderBase#toString throws SOE with the following pattern:
```
	at java.base/java.lang.String.valueOf(String.java:2951)
	at java.base/java.util.concurrent.atomic.AtomicReference.toString(AtomicReference.java:275)
	at java.base/java.lang.String.valueOf(String.java:2951)
	at java.base/java.lang.StringBuilder.append(StringBuilder.java:168)
	at java.base/java.util.EventObject.toString(EventObject.java:76)
	at java.base/java.lang.String.valueOf(String.java:2951)
```
`toString` method is invoked, for example, in tests, when log is printed (e.g. by mockito when it expects something to happen)

GitOrigin-RevId: c9ebec199973f3845f8667345d001e1d271ac9e0