Add option for filtering system nodes
New Option:
===========
Add a filter action button to the component tree tool window header.
The action will control the settings for filtering out system nodes.

Add this option to start and refresh messages such that the agent
can filter out system nodes.

Workaround for now:
===================
Currently the ui depends on receiving the top Decor view.
For now: the agent will include this Decor node even if it is a
system node. We are therefore adding code to filter out this node
from the component tree.

Bug: 136022822
Test: Added unit tests
Change-Id: Id169170341a67f916e08b3565bf028129b8fbb56