[AUA] pass the buildModel to usageInfo
Not just the subsidiary androidModel.  This is both because there is a
possibility that the buildModel might be refreshed between generating
the usage and performing the refactoring based on it (at which point the
older androidModel will no longer be valid) and also for uniformity in
preparation for implementing a more table-driven set of refactorings.

Bug: 178092637
Test: existing (no change in functionality)
Change-Id: I7c35a7266f647c68158c3104b4c4bf642f2f4b5b