[AUA] anonymous BooleanPropertyMove UsageInfos
Instead of an abstract class and named subclasses for UsageInfo
instances corresponding to property moves, we can implement in a more
data-driven way using an inner class of the MoveInfo data, which
contains all the (immutable) information needed to implement the
UsageInfo.  This inner subclass of UsageInfo does need to use the outer
instance to implement the required equality semantics.

Bug: 178092637
Test: existing (no change in functionality)
Change-Id: I19dcab14e0d7a11940157d625ffb8c492cc53f82