[BA] Remove application of BuildAttributionWarningsFilter from analyzers
BuildAttributionWarningsFilter applied in analyzers is not in use at
the moment. Leaving it applied there is only making code more complex
and it will need to be revised anyway once we decide to bring it back.
Removing now. Note that filter is in use at the moment for suppressing
GC warning appearing on UI level, that's why I am not removing it from
everywhere.

Bug: 179137380
Test: existing
Change-Id: I808dddd920d92389b955d7e40e0bb2cb686ee694