Search XML attribute candidates  with AnnotatedElementsSearch
Before this CL, databinding annotations for XML attribute
names are searched directly with JavaAnnotationIndex, which
does not involve much Psi processing. However, when the binding
adapters are declared from third party libraries, the attribute
names are trimmed afrer compiling and we need to call
AnnotatedElementsSearch to find the attribute names.

Fixes: 178122250
Test: Manual Test
Change-Id: I15cc00021bd39e6fc19389d2dea22f8d757938ea