Add ability to add a specific version
We are currently selecting the "best" version of a new dependency
among all the possible versions found in repositories.

With this change a specific version can be specified when calling
Module.addDependenciesWithUiConfirmation in cases where it is
known what the version should be.

It is also possible to specify a major version only or a major and
minor version using the + syntax in version strings.

Bug: 178244644
Test: Unit test added.
Change-Id: I76356117b6818cb37e890fbad29e0d6e4bd1e55e