Remove unnecessary calls to setClockTime
After changing the Compose Animation Preview to use the new timing
controller, we can now remove redundant calls to update the
PreviewAnimationClock. Those calls were done to make sure the clock was
updated on time for the transitions to be notified, e.g. when changing start
and/or end states.

Bug: 177841191
Test: ComposePreviewAnimationManagerTest pass
Change-Id: I76dbcce6815ee7f4907a49e23859fb6a1ee443e1