[Dsl] Support editing existing mutable-collection properties
If we have an implicitly-created mutable collection (from augmenting a
not-yet-present Dsl element using a helper function or augmented
assignment), it is not necessarily legal to change the structure of that
collection by changing the argument list.  For example, some properties
have distinct 1-arg and vararg helper functions, and an attempt to
modify a call which happens to use a 1-arg function to add a new element
would lead to invalid Dsl.  In the case of changed structure, then, we
delete and re-create the element, allowing the backends to make a
suitable choice of representation.

In the case where the structure of the collection is unchanged, though
elements themselves are modified, continue to use the old codepath.
This is a workaround for the fact that the Dsl writer is not yet correct
enough to avoid writing invalid Dsl for some property creation,
particularly those elements of packagingOptions which only have 1-arg
helper functions (and no vararg equivalent); we still cannot in fact
write correct Dsl in general, but at least we don't make the editing
case any worse through this change.

One existing test changes output as a result of this change; the output
remains legal.

Bug: 144280051
Bug: 179483658
Test: added; one test updated
Change-Id: I7be323001625251dd6493d1246880c4cc3e43cd0