Fix persistent progress bar in allocation stage
Occasionally, we get a `FREE_DATA` event with unrecognized tag, which
violates an assertion. Because the error is on a different thread, it
fails silently, and we never remove the progress bar after the
background update finishes.

This change only suppresses the annoying UI symptom. A separate change
will fix the root cause.

Fixes: 179332543
Bug: 180630877
Test: manual
Change-Id: I03f57831c78c40b2c82d9cb26a1bfdc0260185aa