[Dsl] implement RESET through property lookup
Rather than adding a resetting property causing all prior elements of
that property to become HIDDEN, ensure that when looking up a property
that finding a RESET effect on that property has the same effect as
there being none of that property in existence (i.e. returning null from
getPropertyElement()).

Bug: n/a
Test: existing
Change-Id: I7e3efeab6c7086aa52fd7a3d5b70acfd0b5ce9ed