Remove ComposePreviewRunConfigurationProducer in 4.2
When the COMPOSE_PREVIEW_RUN_CONFIGURATION flag is disabled, we add the
ComposePreviewRunConfigurationProducer to the ignored producers list through
RunConfigurationProducerService#addIgnoredProducer when the project opens
(see ComposePreviewRunConfigurationStartupActivity).

However, we started obseving crash reports in Android Studio 4.2 Beta
indicating that ConfigurationContext.findExisting is including this producer
in the producers list even though it's in theory being ignored, since we
disabled the Compose flags in Beta.

While it's not clear why the ignored producers list is being taken into
account, we're removing the ComposePreviewRunConfigurationProducer from the
compose-designer.xml plugin file, so it's not even included in the producers
list that includes the ignored producers.

This change is intended to be merged in 4.2 only. For Arctic Fox, we're
planning to enable the Compose flags and the producer will be enabled.

Bug: 177051695
Test: N/A
Change-Id: I65f6d57924fa57c32b3a5d6d758552773caae66f