Support iterables with references
When extra elements exists in a List/Array on the device, we
will show an extra line at the end of the elements in the
properties panel with a "Show more" link.
Pressing the link will attempt to double the number of elements
in the array.

If we are expanding a reference to another parameter, that other
parameter will also be expanded.

Bug: 180952854
Test: Added to AppInspectionPropertiesProviderTest
Change-Id: Iacd25708f8da981f6a13bb11d1e62a6d30e4adb5