Add build & refresh action to Compose Preview empty panel
When a successful build was not performed yet for a project, we show an
empty panel saying "A successful build is needed before the preview can be
displayed". According to IntelliJ guidelines, we need to provide an action
to fix the issue whenever applicable. Therefore, we provide the "Build &
Refresh" action.

Bug: 180913190
Test: Manual
Change-Id: I2870847a60f5944ddede7b05e8a16d434d335da6
(cherry picked from commit cdbe7a33b27671fb8b33fde258b2f2dc29d0726a)