Add more restrictive file filter when importing UTP results
The previous filter uses "*.pb" but we now have a few other .pb files
that cannot be imported. Thus we make it more restrictive and filter the
name "test-result.pb".

Also add support for parsing text proto. This is because the file filter
does not work very well on Mac and users are still free to select
whatever files other than "test-result.pb".

Bug: 181997773
Test: manual
Change-Id: I82da19ac5bf6e07f7437d45795175d891a06a55c