[BA] Push AGP required for CC version to 7.0.0
AGP supported configuration cache from 4.2 but not all plugins,
e.g. lint supports it only from 7.0.0.
So this CL moves required version to when it becomes fully supported
which is more accurate.
Besides that it helps to eliminate an analyzer state when AGP
is already supported but does not provide any data on other plugins
(which was added in 7.0.0-alpha10).

Bug: 181864876
Test: existing updated
Change-Id: I1f9562bb6f5f6a08db264c18f2f977815de1bdc5