Fixes IDEA-265504: Android: KMPP gradle project with `includeAndroidModules = false` requires Android SDK
Fix invalid assumption that `local.properties` file only exists in Android Gradle Projects (this can also be a KMPP project).
In IDEA: if android SDK is not configured globally, neither it is configured for a project, let user to continue. This might be KMPP project which also uses `local.properties` file.

GitOrigin-RevId: 1d4074ad60f68ac3a66992e2b4856a203b769b01