[BA] Stop using toggleable actions for ui filter items
Platform implementation of toggleable actions in popup menu
messes up with icons which results in not showing the custom
icon we provide for the action. At the same time this default
implementation does not properly support updating the icon
without closing the popup.
Thus it is easier to re-implement desired toggleable behaviour
here since it is quite a trivial change.

Bug: 181212492
Test: manually verified
Change-Id: I3675bb4a17f7d38132a93c52dd0beecdcf1b8ce1