Fix constraint line painting past the arrow in DrawConnection
This fixes a case when for widgets constrained to the parent, the
connection would draw past the parent's bounds. This is
normally not noticeable on XML since the painting is bound by the bounds
of the parent itself (ConstraintLayout view). But for Compose preview,
that's not the case, and the issue is noticeable.

Removed some resulting redundant code.

Fixes: 184092387
Test: n/a, ui tweaks
Change-Id: I739dfd0c4738ca383058f03d9610f44dc6b66d43