Move modification tracker updator logic off the EDT
...since it takes time to iterate all transitive dependents.

Also, there's a small optimization for the use case when the user is
actively editing one manifest file of module with enormous transitive
dependents, and it's very likely a backlog of same requests are created
and pending there. Thus we can just leave one pending request and cancel the
others. (E.g. pending TaskA, Task A, TaskA, TaskB, TaskB -> pending
Task A, pending Task B).

In order to make it simple, we only cache the last request,
and cancel it, if the same new request comes and the old one is still
in progress.

Please note, another issue that enormous threads are depleted when
refreshing merged manifests for large size projects is not addressed
in this cl.

Bug: 184067420
Test: exisiting
Change-Id: Ie94e197c38246aff6c6bad01720c38aea85d2589