Convert //tools/adt/idea/nav/safeargs:testdeps_repo to manifest format
This change adds the `use_zip = False` attribute to
//tools/adt/idea/nav/safeargs:testdeps_repo and updates the accompanying
test suite to use RepoLinker to link the Maven repository rather than
unzip it.

This particular instance of maven_repo was chosen because its single
reverse dependency (go/paste/6732937207742464) makes it a
straightforward maven_repo instance to migrate and demonstrate
RepoLinker integration with IdeaTestSuiteBase.

Bug: 148081564
Test: Existing tests pass
Change-Id: I6da90454579241568022b0cd0d6abfcc5fc6511e