[AUA] Don't reuse a disposed UsageView
Cancelling the preview causes the usage view to be disposed, but does
not otherwise invalidate the Refactoring Processor (by design, in the
Tool Window interaction).  The same Refactoring Processor may be
requested to preview its usages again, meaning that we need to be
careful not to reuse a disposed usageView.

Fixes: 184583482
Test: manual (both feature flag states)
Change-Id: I7c3ad261d1ae12b86463fab559dcf442e888f7ec