Set Compose toolbar DropDownActions icon style to disabled
In order to make it clear for the users that they shouldn't trigger any
actions while the preview is refreshing, we disable both the toolbars of
each individual previews and the file-level toolbar. However, the style of
the two DropDownActions does not change when these actions get disabled, as
IntelliJ will update the style (not the actual enable state) to enabled and
the buttons will only get grayed out when clicking on the actions. Since we
want the users to immediately know these actions are disabled while
refreshing the previews, without needing to click on them, we change the
style of these action icons to disabled when the actions themselves are
disabled.

Bug: 183935976
Test: Manual
Change-Id: Ia1ce62bae63f3104e9d7a89796bbd9f02b3d4b73