[AUA] Don't show checkboxes for individual upgrade steps
Since they're not ever user-togglable independently of the principal
MANDATORY_CODEPENDENT step, they are essentially only used for two
purposes: internal book-keeping convenience (being able to read the
enabled state of a processor straight out of the model) and
layout (ensuring that the labels for the leaf steps all have the same
vertical alignment).  Insert a magic amount of space to preserve the
layout, and hide the checkbox.

Fixes: 184749536
Test: manual
Change-Id: Ia2a79b1f01e0b081ab249eb6fa3ee12007cf1b29