Replace modification tracker in AndroidResolveScopeEnlarger
There exists a case in production, on startup, when AndroidModuleModel
is yet to be set, that agp flags are unavailable. This can leave android
R class scope in incorrect state if a gradle properties was set. When
AndroidModuleModel is set, sync is skipped, and
ProjectRootModificationTracker does not increment. This case is fixed by
using ProjectSyncModificationTracker, which increments on all sync
results.

Fixes: 175326389
Test: Existing LightClassesTest.kt
Change-Id: I33ed32b5eeaf7666f5d082e963331ddb06c2f664