[AUA] Display validation and status message
This might not yet be the final way of displaying the input validation
message: there's an argument that (according to guidelines) it should be
displayed in a tooltip by the ComboBox rather than as an explanation
near the action buttons.  However, it is a useful change in its own
right as it enables us also to display messages relating to the presence
of uncommitted build files and/or unfindable AGP declarations.

Bug: 184638364
Bug: 184752090
Bug: 184865808
Bug: 185179657
Test: manual
Change-Id: I600918ebcdda0bef374102dfd4a19e5a28b4acc0