Add a module template for AndroidX Macrobenchmark
Macrobenchmarking is a planned release as part of androidx.benchmark,
which handles startup benchmarks and profiling. Due to the nature of
macrobenchmarks, they need to be associated with an app module, and an
intent action, but linking these two modules typically requires a lot of
setup.

The hope is that we can ease the setup process for macrobenchmarks using
this wizard.

Fixes: 181673084
Test: ChooseModuleTypeStepTest, NewKtsModuleTest
Test: //tools/adt/idea/android-uitests:AddBenchmarkModuleTest
Change-Id: I0a092e8b630bc9d79c58fa30c2060266fdeed0ad