IDEA-267508: Android: Gradle import fails with AE: syncSucceeded should not be invoked when sync issues present
IDEA will invoke `ProjectDataImportListener#onImportFinished` if gradle returns "SUCCESS". But this is not exactly what Android plugin expects: there might be sync issues like missing platforms - they are reported as warnings and do not fail the sync (in terms of IDEA sync), but the whole operation should be reported as `GradleSyncState#syncFailed` (not `syncSucceeded`)

Fix is to check any reported sync issues in ProjectDataImportListener#onImportFinished before invoking either `syncFailed` or `syncSucceeded`. This is the same logic that is used in AOSP/4.1.1 `PostSyncProjectSetup#notifySyncFinished`.

GitOrigin-RevId: 629a5c0068e32e7a2974d80e3dc172bb46875010