[lint] Compute hierarchy from qualified configuration closures
It is legal, if slightly unusual, to use configuration closures on
qualified expressions rather than unqualified ones.  If we do, we need
to take the qualifiers into account to consider the place in the
hierarchy for the properties being considered.

This does not yet handle the general case of the properties themselves
being qualified.

Bug: n/a
Test: existing, added
Change-Id: Iaff7dcb052fddc17b8c63013e952c25e65845fb8