[lint] Compute hierarchy from qualified statements
If the invoked expression or the assigned property is qualified, the
qualifiers also need to be taken into account when computing the
property's parent and grandparent.  We treat the hierarchy as linear,
treating the property/invoked expression qualifiers as disjoint from the
configuration closure methods -- though particularly eccentric
expressions could (ab)use outer scope implicit receivers to invalidate
this assumption.  This allows us to simply add the closure names found
in the previous change to the names from the expression itself, and read
off the property, its parentName and its parentParentName.  We also add
some additional null checks to make sure not to call detector functions
with invalid arguments.

This allows simplification of the previous special-case for handling
top-level statements, and generalization to all visited properties from
this visitor.

Bug: n/a
Test: added
Change-Id: I8018a81b0bc5a5da63e293b9d77ee3415befcf41