Remove usage of experimental kotlin.time
This is a hotfix for binary compatibility with Kotlin plugin 1.5.0.

The problem is that kotlin.time.Duration is an inline class, and
the binary representation for inline classes is not stable yet
(and e.g. has changed in Kotlin 1.5). This causes binary compatibility
isssues for users who upgrade to a new version of the Kotlin plugin
in the IDE.

(Unfortunately there are no automated checks yet to enforce that we
avoid using experimental Kotlin stdlib APIs in the IDE.)

Bug: relates to b/186464956
Change-Id: Ic79d395f313ef4a498d596d9b889889543315a21
(cherry picked from commit 824b04d2942b33753d390fe8fac4334330d0630a)