[Dsl] Small cleanup of ModelMapCollector
Preparation for more extensive work to install some amount of AGP/Gradle
version-awareness in the Dsl model and parser layers.  In particular
make sure that the elements of data[] are accessed only once, so that if
we have to play games with varying-sized arrays to incorporate more data
that can be handled with only a small amount of pain.

Bug: 186623625
Test: existing (no change in functionality)
Change-Id: Ib15654377fba7c65466562ce9019ebe9335f4424