Show canceled status icon when test run is cancelled
There were two issues which this patch fixes.

1. DdmlibTestRunListenerAdapter notify AndroidTestSuiteView
that a given device is scheduled to be run from its constructor.
If there is an error in deployment tasks or a user cancels the
deployment before AndroidTestApplicationLaunchTask is executed
the cancel for the device is not be notified to
AndroidTestSuiteView.

2. AggregationRow class in AndroidTestResultsTableView doesn't
consume AndroidTestSuiteResult information from the listener.

This patch fixes them by notifying cancellation accordingly
from DdmlibTestRunListenerAdapter and let AggregationRow
refers to AndroidTestSuiteResult.

Bug: 185250931
Test: unit tests and tested manually
Change-Id: I90050773b324d04bd6695eaca004522abfeac5ee