Add support for exception message server flags
With this change, a server-side flag can decide to
include exception message with the exception report. A separate flag
is used to decide to include full exception stack, not just inner-most
exception.

The change prepares functionality for log collection. This functionality
is currently disabled (ExceptionDataCollection.LOG_COLLECTION_ENABLED
is set to false).

Fixes: 185824387
Test: Added ExceptionDataCollectionTest
Change-Id: I56bbbdde81db4628b4a53429608466cff91c853b