Select transition current state when subscribing animation
We were selecting arbitrary states (more specifically, the 0th and 1st
states from the states list) of transition animations being inspected in the
animation preview. That ensured the composable would animate from one state
to the other when user presses play.

In practice, that does not work as expected, and it would be a much better
experience to display the currently selected state of each transition
animation. This way, when users open the animation preview, they will see
the composable in its "resting" state. They can then play around with the
states of each transition to modify the animation and investigate whatever
they want to.

Bug: 187401009
Test: Manual by opening animation preview with Shrine app
Change-Id: I8481e1f1e1f3b2a12447681ead54bf7883a1a29a
(cherry picked from commit 2e0429df48f6eaac100bf834f230a20106175990)