Fix freeze when stopping CPU recording on pre-O devices
For pre-O CPU recording, ddmlib only supports one handler so when
multiple devices are connected, each of them will register their
own handler, overwriting each other. This resuls in freezes when
the STOP_TRACE command waits on a CountDownLatch that is no
longer in scope.

This CL makes LegacyCpuProfilingHandler a singleton that manages
the device to metadata mapping properly.

Fixes: 177612469
Test: added
Change-Id: Ib743bee1a7bb9aa3a87d1bc0e02ee3064dda767d